requirements.txt
django == 5.0.7
python-decouple==3.8
pillow
djangorestframework